new: Rabbit Controller new: Rabbit Form Factory new: Rabbit Form classpath added into config
change: Rabbit Field Factory now accept attributes to apply under construct
change: Initialize call of Rabbit Field is called in Rabbit Form library instead of a self call, now this method is callend only when the Field is fully configured
change: Rabbit Field loadAssets function is removed
change: Fields using loadAssets is now using initialize instead

remove: Main controller is removed, replaced by Rabbit Controller
remove: All files that are not from Rabbit Forms is removed from project (like CI config files)

bug fix: warning when delete file that not exists

+++ rabbit-forms/application/rabbit-forms/lib/Rabbit/Field/File.php
@@ -141,10 +141,12 @@ class Rabbit_Field_File extends Rabbit_Field
             $id
         ))->row_array();
 
-        $path = $this->baseFilePath() . $id . '_' . $data[$this->getName()];
+        if(isset($data[$this->getName()])) {
+            $path = $this->baseFilePath() . $id . '_' . $data[$this->getName()];
 
-        if(file_exists($path)) {
-            unlink($path);
+            if(file_exists($path)) {
+                unlink($path);
+            }
         }
     }
